docs: Remove template "itemizedlist/listitem" from man.xsl
This seems to have been basically taken from the manpages/lists.xls
from the docbook-xsl stylesheets. But it references a variable list-indent
that older versions of docbook-xsl (e.g. 1.69) do not provide.
This makes the manpage build break on older systems. Removing
the definition lets the build succeed, using the system-definition
of the itemizedlist/listitem.

The diff between the docbook's (version 1.75.1) definition of
itemizedlist/listitem and the definition in our man.xls is this:

-- with this patch
-- without this patch
@@ -53,5 +53,7 @@
   <!-- * seems to require the extra space. -->
   <xsl:call-template name="roff-if-end"/>
   <xsl:apply-templates/>
-  <xsl:text>.RE&#10;</xsl:text>
+  <xsl:if test=" following-sibling::listitem">
+        <xsl:text>&#10;.RE&#10;</xsl:text>
+  </xsl:if>
 </xsl:template>

I.e. the version of man.xsl made insertion if ".RE" conditional.
I hope this does not break anything severely.
The diff for e.g. the resulting winbindd.8 manpage is this:

--- with this patch
+++ witout this patch:
@@ -375,7 +375,6 @@

 \m[blue]\fBwinbind: rpc only\fR\m[]
 Setting this parameter forces winbindd to use RPC instead of LDAP to retrieve information from Domain Controllers\&.
-.RE
 .SH "EXAMPLE SETUP"
 .PP
 To setup winbindd for user and group lookups plus authentication from a domain controller use something like the following setup\&. This was tested on an early Red Hat Linux box\&.

Cheers