s4-dsdb Fix generation of rootDSE domainControllerFunctionality
The issue here is that when the samdb calls were removed, this call
relied on going back to the top of the module stack, so as to re-enter
the rootDSE search handler.  It makes more sense to check the @ROOTDSE
record directly, and therefore not to invoke the search() handler
during the init.

Andrew Bartlett