We default to non-deterministic ip now where ips are "sticky" and dont change too much. This means we can simplify the way we add ips significantly and stop trying to move them.
We also check if the node already hosts the ip, in which case we used to return an error. Instead just print an error string but return 0, ok.
This makes it easier to script, and works around broken scripts.

CQ1021034

(This used to be ctdb commit 307e5e95548155a31682dfcb0956834d0c85838e)