s4-python: get ordering right on python paths
we need to have PYTHONARCHDIR first, as that is where we currently
install. That will minimise problems with upgrades from old installs.

We also don't need to add the path twice if PYTHONDIR is the same as
PYTHONARCHDIR (as happens on debian based systems)

Pair-Programmed-With: Andrew Bartlett <abartlet@samba.org>