s4:ldapcmp: cope with range retrivals of multivalued attributes
A Windows Server returns a 'member;range=0-1499' attribute
with the first 1500 values of the 'member' attribute.

The client can do a BASE search on the given object
and ask for the 'member;range=1500-*' attribute.
It needs to loop until the high part of the returned
range is '*'.

metze

Autobuild-User: Stefan Metzmacher <metze@samba.org>
Autobuild-Date: Mon Feb 14 16:26:46 CET 2011 on sn-devel-104