build: created samba-util-common library
this library will contain files that can be built in common between s3
and s4. Both samba-util and samba-util3 depend on it.

Currently this library needs to be marked as allow_undefined_symbols,
in order to allow s4 to build it. We can remove that flag once we have
all the needed symbols in common

Pair-Programmed-With: Andrew Bartlett <abartlet@samba.org>