heimdal_build omit #line statments to allow valgrind to work again
The lex/yacc files were generated on Fedora 14, and have empty
filenames in #line declarations.  I don't know why this is, but it
seems best just to omit the #line statements.

This is what was causing Valgrind on Fedora not to run on Samba
binaries and programs linked to Samba libraries.

Andrew Bartlett

Autobuild-User: Andrew Bartlett <abartlet@samba.org>
Autobuild-Date: Fri Feb 25 11:46:56 CET 2011 on sn-devel-104