s4-dsdb: implemented creation of conflict records
when a record with the same DN gets created on two DCs at the same
time, this creates a replication conflict. To resolve this conflict
one of the DCs needs to create a conflict record, which is a rename of
one of the two DNs, based on which one is newer.

This prevents replication from failing when DCs are temporarily
disconnected and then have conflicts when they next replicate

Pair-Programmed-With: Andrew Bartlett <abartlet@samba.org>

Autobuild-User: Andrew Tridgell <tridge@samba.org>
Autobuild-Date: Mon Apr  4 03:27:07 CEST 2011 on sn-devel-104