s3-selftest Fix use of the 'fake DNS' file
This ensures we don't use netbios before the hosts file and that we do
fill in the fake DNS zone correctly for the way we invoke smbtorture4.
Currently this works because the realm in client.conf is "", if this
changes then this will need to change too (perhaps an additional
entry).

Andrew Bartlett