s3 waf: Mimic the autoconf check for netgrent
The autoconf check for netgrent headers makes sure that the compiler flag used
(-Werror-implicit-function-declaration) is recognized by the compiler before
using it for the test. Mimic this behaviour in the waf checks.
Thanks to gd for noticing this.

Signed-off-by: Günther Deschner <gd@samba.org>

Autobuild-User: Günther Deschner <gd@samba.org>
Autobuild-Date: Mon Apr 18 16:24:15 CEST 2011 on sn-devel-104