s3-test Make samba3.blackbox.net.local.registry.roundtrip reliable
Looking for smb.conf files with find -name \*.conf would find and not
find all sorts of things that were just not valid inputs.  When
restricted to under source3 it then finds no inputs in the top level
build, as it only worked on the generated smb.conf files from 'make
test' environments.

Instead, just run it on the one file we know is valid, and if we want
to run it on more files, then we can make this test run against more
environments, or specific target files.

Andrew Bartlett