vfs_gpfs: Properly notify the offline->online changes
This needs to be a separate module that cooperates with vfs_gpfs. If aio_fork
is used early in the module chain it (correctly) does not propagate the aio ops
down, so vfs_gpfs does not see them. This slim module must come early in the
chain for notifies to work properly.
(cherry picked from commit 3f68b43dd56b8f338a18074741162bf6437ae78d)