s3-events: tevent_internal.h is not a public header
We need to use the "foo.h" form instead of the <foo.h> form for
headers that are not installed publicly. Otherwise when an external
version of tevent is used we won't find the header and the build will
fail.

Note that this creates a structure dependency between the
tevent structures in the external library and the headers in our
source tree. That is not ideal, but is currently OK as the waf build
will only use the external library if it is at least the same version
as the internal tree, which means it will actually be the same
version, as we release the external version from our tree.

We should come up with a better solution, but for now this allows
openchange to build again.

Autobuild-User: Andrew Tridgell <tridge@samba.org>
Autobuild-Date: Tue May 10 16:06:40 CEST 2011 on sn-devel-104
(cherry picked from commit 628fcd111af9b51020191a6ee916896c127c88c1)
(cherry picked from commit 6b919c31974be4167f217b720e975c37ee94d6a3)