selftest: Make knowfail/skip files consistent, always use ^prefix
Except in one case (where we mark printing tests as knownfail), this
has all our regular expressions start with ^, which ensures we don't
accidentially mix up the samba3/samba4 prefix here.

Because of the particular values in these files at the moment, this
should not change the set of tests, but it will help to ensure that
future edits follow the correct pattern.

Andrew Bartlett